simple fix for leveldb format support
moooooooooooooooooore todo in NBT parsing while PE have a really
different structure with Level/Entity/Tile.

+++ src/main/java/cn/nukkit/level/format/LevelProvider.java
@@ -51,9 +51,9 @@ public interface LevelProvider {
 
     String getName();
 
-    int getTime();
+    long getTime();
 
-    void setTime(int value);
+    void setTime(long value);
 
     long getSeed();
